Add check that ensures a candidate gid doesn't exist in the local 'files' db
Previously, userman would assume that there were no groups with gids
about 3000 defined in /etc/group.  This worked fine.  At some recent
point in time, some local group-creation utility started noticing the
ldap groups and groups created with it got gids above 3000.  This
resulted in collisions and unpleasantness.

This patch makes the scripts check that a candidate gid is not already
in use on the local machine.