Add a simple username existence check.
This moves the user loading logic to a separate method and uses this
for both actual user loading and for existence checks. Deliberately
re-use this logic rather than re-implementing an existence check in
order to ensure that user.exists and user(name).in_db will return
the same results.

The primary intention of this method is to replace the need to create
a new user object just to check the in_db flag, which is quite slow
when it fails (I suspect due to the password generation).

Change-Id: I1d179e7dbd6776a11453969864ee7ac08b5b4d5c